[fix] [test] Fix flaky test ReplicatorTest (#22594)
(cherry picked from commit 6fdc0e31bff906446e70965531671389d57e6cda)
(cherry picked from commit 9242f333d9ce112c9f7ce1497380ab5e059429f7)
